feat: refactor invoice creation process to use DTOs and enhance product handling with BigDecimal for monetary values

+++ src/main/java/com/hiagosouza/api/quoted/dtos/invoice/InvoiceRequestDTO.java
@@ -14,8 +14,8 @@ public class InvoiceRequestDTO {
     private String customerDocument;
     private List<InvoiceProductRequestDTO> products;
     private InvoiceStatus invoiceStatus;
-    private Double tax;
-    private Double discount;
+    private BigDecimal tax;
+    private BigDecimal discount;
     private BigDecimal total;
     private BigDecimal subtotal;
     private LocalDateTime createdAt;
